Split up `test_builtins.py` into two files
